--- a/Pipfile
+++ b/Pipfile
@@ -9,7 +9,6 @@ gunicorn = "*"
 marshmallow = "~=2.17.0"
 prometheus_client = "*"
 requests = ">=2.20.0"
-pandas = "*"
 
 [dev-packages]
 pytest = "*"