--- a/custom_parser/tar_gz_csv.py
+++ b/custom_parser/tar_gz_csv.py
@@ -1,17 +1,20 @@
+import json
 import tarfile
 from io import BytesIO
-from itertools import chain
+
 import pandas as pd
 
 
-def csv_parser(file_obj: BytesIO) -> list:
+def csv_parser(file_obj: BytesIO) -> bytes:
     """Extract CSV data from TAR.GZ file.
 
     Extracts all CSV files from a TAR.GZ archive and combines them into a list
     of dictionary objects.
     :param filename: Name of the filename to parse
     """
-    entries = list()
+    # Start data stream
+    yield b'['
+
     with tarfile.open(fileobj=file_obj) as tar:
         for member in tar:
             # Only CSV files are interesting
@@ -20,6 +23,10 @@ def csv_parser(file_obj: BytesIO) -> list:
 
             # Read the CSV
             csv_data = pd.read_csv(tar.extractfile(member))
-            entries = chain(entries, csv_data.to_dict('records'))
 
-    return list(entries)
+            for item in csv_data.itertuples(index=False):
+                out_str = json.dumps(item._asdict())
+                yield f'{out_str},'.encode()
+
+    # End data stream
+    yield b']'