--- a/workers.py
+++ b/workers.py
@@ -44,6 +44,13 @@ def _retryable(method: str, *args, **kwargs) -> requests.Response:
     raise requests.HTTPError('All attempts failed')
 
 
+def _output_gen(source_id, file_obj):
+    yield f'{{"id":{source_id},"data":'.encode()
+    for chunk in custom_parser.parse(file_obj):
+        yield chunk
+    yield b'}'
+
+
 def download_job(
         source_url: str,
         source_id: str,
@@ -83,19 +90,13 @@ def download_job(
 
         # Unpack data and stream it
 
-        # Build the POST data object
-        data = {
-            'id': source_id,
-            'data': custom_parser.parse(file_obj),
-        }
-
         # Pass to next service
         prometheus_metrics.METRICS['posts'].inc()
         try:
             resp = _retryable(
                 'post',
                 f'http://{dest_url}',
-                json=data,
+                data=_output_gen(source_id, file_obj),
                 headers={"x-rh-identity": b64_identity}
             )
             prometheus_metrics.METRICS['post_successes'].inc()
@@ -156,6 +157,7 @@ def download_job(
                 thread.name, source_id, exception
             )
             prometheus_metrics.METRICS['post_errors'].inc()
+        prometheus_metrics.METRICS['post_successes'].inc()
 
         logger.debug('%s: Done, exiting', thread.name)