Current way of parsing the received file is inefficient. This Pr provides means to optimize memory usage and improve time costs per job.

## Means used to improve efficiency

These are the changes I've made along with the reasoning behind.

### Remove tempfile logic [`e6a3822`](https://github.com/ManageIQ/aiops-data-collector/pull/16/commits/e6a38229d49fdf4e5d1c9a3c3857db597d2717c2)

_Red series in the chart_

This is more of a code cleanup. The presumption is that once we receive the file it doesn't make any sense to store it as a temp file, because in containerized environment the `/tmp` partition is a `tmpfs` type - the data are stored in memory nevertheless. Therefore, we won't save much memory if we would dump the data in a file and then do our parsing over this tempfile.

Using a temfile would save us the HTTP response data overhead - a bit less memory due to `tmpfs` optimization on OS level. But it takes more time to process the data - it takes time to store the file and load it again.

### Use chunked stream [`864e93c`](https://github.com/ManageIQ/aiops-data-collector/pull/16/commits/864e93c911f6552ba415700623b2015c5ccb5183)

_Yellow series_

Unfortunately, if we want to parse the `*.csv` we need to load it fully into memory. But when we do so, we don't need to parse it as a whole. Instead we can parse each line as a separate entry. This way we don't need to keep the whole output file in memory and we can instead stream each entry separately in a chunked manner. Our memory usage then lowers to "just":
- whole `*.tar.gz` in memory
- whole `*.csv` in memory
- one entry parsed at a time

This approach greatly lowers memory usage, since it takes much space to store whole file while parsed. Pandas and Numpy are root cause of this issue. We've used them for the parsing instead of the standard `csv` package, because that one doesn't handle well data conversions... This issue is completely avoided with the last optimization:
 
### Avoid parsing to JSON and stream chunked binary data [`f0705eb`](https://github.com/ManageIQ/aiops-data-collector/pull/16/commits/f0705eb56a7ca36bb421d216455ad364fbe69179)

_Green series_

The parsing would be done once more on the AI service side nevertheless. Hence it doesn't make much sense to do it twice just so we can have neater, readable output of the `data-collector`. Since the parsing would consume time and resources in the AI service, the data collector can skip that part and just unpack the file and dump the binary data as a stream. We can chunk the reads of the `*.csv` - limit to a buffer size. The size limit can be used to balance memory spikes to time it takes to finish the job.

Memory and time savings are huge since the parsing is really costly operation and we're about to avoid that.

## Benchmark

To see how each optimization performs, we should do some benchmark. For that I've chose to use the Titanic dataset once more, just layered the data... So I've extended the data set by the same data so I've end up with a file:

Metric | Value
----- | -----
Total size of `titanic_extra_big.tar.gz` when compressed | 66MiB
Inner `titanic.csv` when unpacked | 211MiB
Inner `titanic.csv` entries total | 4436774 lines + 1 header

So as you can see it's really sparse and not much unique. That helps with compression so it shortens the time it takes to pass the file over the network. Also the impact of storing the file locally is lowered as well.

Now, let's see how each of the optimizations performs:

![chart 1](https://user-images.githubusercontent.com/7453394/51982879-7d28d380-2497-11e9-8d28-0e0d1b689d08.png)

Optimization | Memory peak for the same job
---------------- | ------------------
**Master** for comparison | 4433.92MiB
**No temp file** | 5092.35MiB
**Chunked parsed stream** | 1115.14MiB
**Raw binary stream** |  134MiB

It's worth mentioning the data were collected using `docker stats` which have a refresh rate of approx. a second. So the peak you see in the **No temp file** series may have (and probably did) occurred in the **Master** as well, it just wasn't caught.

Optimization | Time to complete a job
---------------- | ------------------
**Master** for comparison | 137s
**No temp file** | 135s
**Chunked parsed stream** | 23s
**Raw binary stream** | 18s

## Conclusion

As shown, chunked stream performs far better than parsing whole file at once. That is a valid observation no matter if we choose to parse the data or just dump it raw.

All these approaches results in the same, consumable output. Due to change in data format, only the last **Raw binary stream** optimization would require a small change in the AI service.

The residual memory increase after a job is completed, is something worth looking at. It seems the garbage collector in Python doesn't work well when the `Thread` objects are nod joined back... This should be addressed in future.