@AparnaKarve 2019-01-31
@tumido This is a really nice explanation of how the commits co-relate to the graph data, and the results are really impressive!

I will test this out soon.

Couple questions  -
- Were the above readings taken on your local machine (with the microservice running from the docker instance)?
- The Time on the x-axis - would that be time in minutes?
- how many times was the POST request made?