@AparnaKarve 2019-02-01
You are right, I should try the test with larger datasets.

I tried the Single Request test about 5 times and 2 out of those 5 times (like I mentioned above) the consumption in master was less than streamed - master was at about 25MiB and streamed was at 30MiB.
Then I had to restart the docker daemon on my machine, and when I resumed my testing, I saw the exact opposite - streamed consumption was less than master. So I repeated the test 3 more times, and this time streamed consumed less in every test.

But if you disregard that first spike for both master and streamed (just to be fair), I think the streamed approach did perform better in terms of time

I do like the streamed implementation a lot, so if it works well with https://github.com/ManageIQ/aiops-insights-clustering/pull/29 (which I will test next), I think this should be good to go!