@tumido 2019-02-01
@AparnaKarve Sounds cool! :wink: 

I don't know... the behavior for a single request on your machine is kinda weird. :smile: Let's see what results you're gonna get, when you run some more expensive test. But I'm pretty sure that it's gonna behave similar to my observations..

The insights-clustering PR brings no change to their workflow, it just adopts to these changes, so I don't expect any delays on that side. If we decide we want to go this way and merge this, the only change we have to keep in mind is that the API spec is changed for the AI service... Other than that, we should be fine. :slightly_smiling_face: