@AparnaKarve 2019-02-02
@tumido I am currently testing with -
`'{"url": "https://github.com/AparnaKarve/datasets/blob/master/big.tar.gz?raw=true"}'`

It's a fairly large dataset - size: 15.5 MB and 65,500 rows

the results are better for master I think -- but need a second pair of eyes and another machine to confirm.

Meanwhile,  I also tested ManageIQ/aiops-insights-clustering#29, and that looks good!