@tumido 2019-02-04
Hm, the results are significantly for me, however the data set you're using is flawed. It contains 2 `CSV` files, and since the latest "raw streamed" commit is set to parse only the first CSV file found, it makes the results incomparable...

Try my dataset, which is still by magnitude larger than the one you've used:

https://drive.google.com/file/d/1aAqMyRBA7_0f6jkUJPsnY_t0q2sk0p1C/view?usp=sharing

Also, more and more multiple tests I run, I feel like I need to address the issue with cleanup after the workers.. :/ That has to be fixed, it seems the memory is not cleaned, when the worker exits... Never. Which is bad.