@tumido 2019-02-21
@AparnaKarve what do you think about this PR, can we merge it (after the demo though...)? So I can base my multi-worker on it? :) Any objections? :cake: