@AparnaKarve 2019-02-22
@tumido I tested this again today a couple times using your dataset 

cAdvisor readings -

|Scenario | Master  | Streamed |
| -----|------------- | ------------- |
| Idle state |73.10  | 71.87  |
| Single Run | 76.04  | 74.29  |
| Multiple Runs(4) | 79.18  | 80.18  |
| Multiple Runs(6) | 82.49  | 80.07  |
| Multiple Runs(12) | 87.21  | 80.14  |
| Multiple Runs(20) | 89.13  | 80.02  |


Memory never gets restored back to it's original consumption number for both cases, which I believe even you noted.

From what I can see, the streamed seems to be performing better with more number of runs -- it's staying consistent at ~80MiB 

If we average it out over several runs, clearly streamed would do better than master.

Good to go from my side!! :)