@tumido 2019-02-22
@AparnaKarve  I don't get how you can have such low RAM readings on master... I get nowhere near that. :/ Like at all. Are you sure you're building the image the proper way?

```bash
$ git checkout master
$ s2i build -c . centos/python-36-centos7 aiops-data-collector:master
```

I mean.. it's really interesting that you're deployment gets so low in consumption. I can't reproduce it either locally or in OpenShift.