@AparnaKarve 2019-02-22
@tumido I checked out both `master` and `streamed_data` branches and did this for each branch respectively -
```
docker build -t <my_docker_id>/aiops-data-collector_master -f Dockerfile .
docker build -t <my_docker_id>/aiops-data-collector_streamed -f Dockerfile .
```

I was surprised at the low numbers too, but it is what it is.
When I was testing a couple days ago, the numbers were not this low, I remember.

Also I left the docker instances ON overnight, and the memory consumption did come down today quite a bit -
Master - 76.40
Streamed - 74.61