@tumido 2019-02-22
Hm.. weird.. we don't have a `Dockerfile` in the repo. What are using to build it against then? Please use the `s2i` so the workflow is the same as for OpenShift.