@tumido 2019-02-22
I mean.. if your dockerfile features a better setup for the master.. Let's use it! :wink: :wine_glass: