@AparnaKarve 2019-02-22
We did have a Dockerfile at some point, right?
It was removed in https://github.com/ManageIQ/aiops-data-collector/pull/11 - but that is what I am using.

I will be retesting with `s2i` today