@tumido 2019-02-22
@AparnaKarve Hm.. that `Dockerfile` is dated. It also requires `requirements.txt` which we no longer keep in the repo, so you've probably had the images using wrong and incomplete dependencies... Did you check on the outputs during your tests? Well.. anyways, let's test it against `s2i` builds which are up-to-date. I think that would give you the same results as it did for me.

Also note, that you have to have the parser set to `CSV` in order to kick in. Otherwise the pandas won't run at all... That's maybe the issue. The Dockerfile with dated requirements (missing pandas and stuff) and no parsing happening (otherwise you'd get traceback because of missing pandas)... no wonder the results are nearly the same for you.

One mystery solved. :slightly_smiling_face: Let's try with `s2i`  and we'll see. :slightly_smiling_face: